added pylint/mypy complaints
